moved _preprocess_dims, _stack_input_if_needed, _preprocess_weights from deterministic.py to utils.py
