+++ xskillscore/core/utils.py
@@ -5,6 +5,99 @@ from xhistogram.xarray import histogram as xhist
 __all__ = ['histogram']
 
 
+def _preprocess_dims(dim):
+    """Preprocesses dimensions to prep for stacking.
+
+    Parameters
+    ----------
+    dim : str, list
+        The dimension(s) to apply the function along.
+    """
+    if isinstance(dim, str):
+        dim = [dim]
+    axis = tuple(range(-1, -len(dim) - 1, -1))
+    return dim, axis
+
+
+def _stack_input_if_needed(a, b, dim, weights):
+    """
+    Stack input arrays a, b if needed.
+    Adapt dim and weights accordingly.
+
+    Parameters
+    ----------
+    a : xarray.Dataset or xarray.DataArray
+        Labeled array(s) over which to apply the function.
+    b : xarray.Dataset or xarray.DataArray
+        Labeled array(s) over which to apply the function.
+    dim : list
+        The dimension(s) to apply the metric along.
+    weights : xarray.Dataset or xarray.DataArray or None
+        Weights matching dimensions of ``dim`` to apply during the function.
+
+    Returns
+    -------
+    a : xarray.Dataset or xarray.DataArray stacked with new_dim
+        Labeled array(s) over which to apply the function.
+    b : xarray.Dataset or xarray.DataArray stacked with new_dim
+        Labeled array(s) over which to apply the function.
+    new_dim : str
+        The dimension(s) to apply the metric along.
+    weights : xarray.Dataset or xarray.DataArray stacked with new_dim or None
+        Weights matching dimensions of ``dim`` to apply during the function.
+    """
+    if len(dim) > 1:
+        new_dim = '_'.join(dim)
+        a = a.stack(**{new_dim: dim})
+        b = b.stack(**{new_dim: dim})
+        if weights is not None:
+            weights = weights.stack(**{new_dim: dim})
+    else:
+        new_dim = dim[0]
+    return a, b, new_dim, weights
+
+
+def _preprocess_weights(a, dim, new_dim, weights):
+    """Preprocesses weights array to prepare for numpy computation.
+
+    Parameters
+    ----------
+    a : xarray.Dataset or xarray.DataArray
+        One of the arrays over which the function will be applied.
+    dim : str, list
+        The original dimension(s) to apply the function along.
+    new_dim : str
+        The newly named dimension after running ``_preprocess_dims``
+    weights : xarray.Dataset or xarray.DataArray or None
+        Weights to apply to function, matching the dimension size of
+        ``new_dim``.
+    """
+    if weights is None:
+        return None
+    else:
+        # Throw error if there are negative weights.
+        if weights.min() < 0:
+            raise ValueError(
+                'Weights has a minimum below 0. Please submit a weights array '
+                'of positive numbers.'
+            )
+        # Scale weights to vary from 0 to 1.
+        weights = weights / weights.max()
+        # Check that the weights array has the same size
+        # dimension(s) as those being applied over.
+        drop_dims = {k: 0 for k in a.dims if k not in new_dim}
+        if dict(weights.sizes) != dict(a.isel(drop_dims).sizes):
+            raise ValueError(
+                f'weights dimension(s) {dim} of size {dict(weights.sizes)} '
+                f"does not match DataArray's size "
+                f'{dict(a.isel(drop_dims).sizes)}'
+            )
+        if dict(weights.sizes) != dict(a.sizes):
+            # Broadcast weights to full size of main object.
+            _, weights = xr.broadcast(a, weights)
+        return weights
+    
+    
 def histogram(*args, bins=None, bin_names=None, **kwargs):
     """Wrapper on xhistogram to deal with Datasets appropriately
     """
